♻️ Move stmt.go to internal/wrap
