+++ conn.go
@@ -106,7 +106,7 @@ func (c *connWrapper) Prepare(query string) (driver.Stmt, error) {
 	if attr != nil {
 		lg = lg.With(*attr)
 	}
-	return wrapStmt(origStmt, lg), nil
+	return wrap.WrapStmt(origStmt, lg), nil
 }
 
 // IsValid implements driver.Validator.
@@ -233,7 +233,7 @@ func (c *connWithContextWrapper) PrepareContext(ctx context.Context, query strin
 	if attr != nil {
 		lg = lg.With(*attr)
 	}
-	return wrapStmt(stmt, lg), nil
+	return wrap.WrapStmt(stmt, lg), nil
 }
 
 // BeginTx implements driver.ConnBeginTx.
